Split take-first-screenshot
