+++ src/take-first-screenshot.h
@@ -1,0 +1,31 @@
+/* take-first-screenshot.h
+ *
+ * Copyright 2025 Kelvin
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+#pragma once
+
+#include <libportal/portal.h>
+
+#include "kasasa-window-private.h"
+
+G_BEGIN_DECLS
+
+void take_first_screenshot (KasasaWindow *window);
+
+G_END_DECLS
